Bring back zsh-autoenv
Tried direnv. Nice idea, but seriosuly running new process just to
collect env vars and calling that lightweight is kind weird to me.

Not to say that it pollutes output with `loading... +FOO` bla bla...
That would be nice in debug mode, but on regular basis not for me.

Finally due to it's architecture I can't run `chruby` for example
to change ruby upon dir entry... So... it's not for me at least right
now...